Homework 05.07.2021 (redone)
